+++ src/main/java/bowling/domain/pitchings/LastFramePitchings.java
@@ -3,19 +3,9 @@ package bowling.domain.pitchings;
 import bowling.domain.KnockDownPins;
 import bowling.domain.Pitching;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.stream.Stream;
-
-public class LastFramePitchings implements Pitchings {
+public class LastFramePitchings extends Pitchings {
     private static final String INVALID_PITCHING_CHANCE_ERR_MSG = "투구 기회가 없습니다.";
     private static final int LAST_FRAME_MAX_PITCHING_SIZE = 3;
-    private final List<Pitching> value;
-
-    public LastFramePitchings() {
-        this.value = new ArrayList<>();
-    }
 
     public static LastFramePitchings getInstance() {
         return new LastFramePitchings();
@@ -41,20 +31,8 @@ public class LastFramePitchings implements Pitchings {
         throw new IllegalStateException(INVALID_PITCHING_CHANCE_ERR_MSG);
     }
 
-    private void setFirstPitching(KnockDownPins knockDownPins) {
-        Pitching pitching = Pitching.getPitching(knockDownPins);
-        value.add(pitching);
-    }
-
     private boolean isSecondPitchingSequence() {
         return value.size() == 1;
-    }
-
-    private void setSecondPitching(KnockDownPins knockDownPins) {
-        int lastIndex = value.size() - 1;
-        Pitching previousPitching = value.get(lastIndex);
-        Pitching pitching = Pitching.getPitching(knockDownPins, previousPitching);
-        value.add(pitching);
     }
 
     private boolean hasThirdPitchingChance() {
@@ -92,43 +70,5 @@ public class LastFramePitchings implements Pitchings {
 
     private boolean hasPitchingChance() {
         return value.size() < LAST_FRAME_MAX_PITCHING_SIZE - 1;
-    }
-
-    @Override
-    public Iterator<Pitching> iterator() {
-        return value.iterator();
-    }
-
-    @Override
-    public Stream<Pitching> stream() {
-        return value.stream();
-    }
-
-    @Override
-    public boolean contains(Pitching pitching) {
-        return value.contains(pitching);
-    }
-
-    @Override
-    public int getTotalScore() {
-        return value.stream()
-                .mapToInt(Pitching::getScore)
-                .sum();
-    }
-
-    @Override
-    public Pitching getNextPitching() {
-        if (value.isEmpty()) {
-            return null;
-        }
-        return value.get(0);
-    }
-
-    @Override
-    public Pitching getNextAndNextPitching() {
-        if (value.size() < 2) {
-            return null;
-        }
-        return value.get(1);
     }
 }
