+++ src/main/java/bowling/domain/pitchings/NormalFramePitchings.java
@@ -3,18 +3,8 @@ package bowling.domain.pitchings;
 import bowling.domain.KnockDownPins;
 import bowling.domain.Pitching;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.stream.Stream;
-
-public class NormalFramePitchings implements Pitchings {
+public class NormalFramePitchings extends Pitchings {
     private static final int NORMAL_FRAME_MAX_PITCHING_SIZE = 2;
-    private final List<Pitching> value;
-
-    public NormalFramePitchings() {
-        this.value = new ArrayList<>();
-    }
 
     public static NormalFramePitchings getInstance() {
         return new NormalFramePitchings();
@@ -28,18 +18,6 @@ public class NormalFramePitchings implements Pitchings {
         }
 
         setSecondPitching(knockDownPins);
-    }
-
-    private void setFirstPitching(KnockDownPins knockDownPins) {
-        Pitching pitching = Pitching.getPitching(knockDownPins);
-        value.add(pitching);
-    }
-
-    private void setSecondPitching(KnockDownPins knockDownPins) {
-        int lastIndex = value.size() - 1;
-        Pitching previousPitching = value.get(lastIndex);
-        Pitching pitching = Pitching.getPitching(knockDownPins, previousPitching);
-        value.add(pitching);
     }
 
     @Override
@@ -58,44 +36,5 @@ public class NormalFramePitchings implements Pitchings {
     private boolean isFirstPitchingStrike() {
         Pitching firstPitching = value.get(0);
         return firstPitching == Pitching.STRIKE;
-    }
-
-    @Override
-    public Iterator<Pitching> iterator() {
-        return value.iterator();
-    }
-
-    @Override
-    public Stream<Pitching> stream() {
-        return value.stream();
-    }
-
-    @Override
-    public boolean contains(Pitching pitching) {
-        return value.contains(pitching);
-    }
-
-    @Override
-    public int getTotalScore() {
-        return value.stream()
-                .mapToInt(Pitching::getScore)
-                .sum();
-    }
-
-    @Override
-    public Pitching getNextPitching() {
-        if (value.isEmpty()) {
-            return null;
-        }
-        return value.get(0);
-    }
-
-    @Override
-    public Pitching getNextAndNextPitching() {
-        if (value.size() < 2) {
-            return null;
-        }
-
-        return value.get(1);
     }
 }
